+++ src/types/new_types/mod.rs
@@ -1,6 +1,7 @@
 mod accounts;
 mod factor_instances;
 mod is_securified_entity;
+mod key_space;
 mod securified_account;
 mod securified_accounts;
 mod securified_entity;
@@ -12,6 +13,7 @@ mod veci;
 pub use accounts::*;
 pub use factor_instances::*;
 pub use is_securified_entity::*;
+pub use key_space::*;
 pub use securified_account::*;
 pub use securified_accounts::*;
 pub use securified_entity::*;
